src/requirements.txt
Pillow
pygame
easygui